requirements.txt
streamlit==1.45.0
pandas
numpy
requests
plotly
thefuzz
matplotlib.pyplot

